+++ src/rendering/line_iter.rs
@@ -12,13 +12,6 @@ use crate::{
 use az::{SaturatingAs, SaturatingCast};
 use embedded_graphics::{pixelcolor::Rgb888, prelude::PixelColor};
 
-#[cfg(feature = "ansi")]
-use super::ansi::try_parse_sgr;
-#[cfg(feature = "ansi")]
-use ansi_parser::AnsiSequence;
-#[cfg(feature = "ansi")]
-use as_slice::AsSlice;
-
 /// Parser to break down a line into primitive elements used by measurement and rendering.
 #[derive(Debug)]
 #[must_use]
@@ -123,8 +116,7 @@ where
                     break 'lookahead;
                 }
 
-                #[cfg(feature = "ansi")]
-                Some(Token::EscapeSequence(_)) => {}
+                Some(Token::ChangeTextStyle(_)) | Some(Token::MoveCursor { .. }) => {}
 
                 _ => break 'lookahead,
             }
@@ -188,16 +180,11 @@ where
                 Some(Token::Whitespace(n, _)) => spaces.consume(n).saturating_as(),
                 Some(Token::Tab) => cursor.next_tab_width().saturating_as(),
 
-                #[cfg(feature = "ansi")]
-                Some(Token::EscapeSequence(AnsiSequence::CursorForward(by))) => by.saturating_as(),
-
-                #[cfg(feature = "ansi")]
-                Some(Token::EscapeSequence(AnsiSequence::CursorBackward(by))) => {
-                    -by.saturating_as::<i32>()
+                Some(Token::MoveCursor { chars, .. }) => {
+                    chars * handler.measure(" ").saturating_as::<i32>()
                 }
 
-                #[cfg(feature = "ansi")]
-                Some(Token::EscapeSequence(_)) => 0,
+                Some(Token::ChangeTextStyle(_)) => 0,
 
                 _ => return false,
             };
@@ -379,47 +366,39 @@ where
                     }
                 }
 
-                #[cfg(feature = "ansi")]
-                Token::EscapeSequence(seq) => {
-                    match seq {
-                        AnsiSequence::SetGraphicsMode(vec) => {
-                            if let Some(sgr) = try_parse_sgr(vec.as_slice()) {
-                                handler.change_text_style(sgr.into())?;
+                // Cursor movement can't rely on the text, as it's permitted
+                // to move the cursor outside of the current line.
+                // Example:
+                // (| denotes the cursor, [ and ] are the limits of the line):
+                // [Some text|    ]
+                // Cursor forward 2 characters
+                // [Some text  |  ]
+                Token::MoveCursor {
+                    chars,
+                    draw_background: true,
+                } => {
+                    let delta = chars * handler.measure(" ").saturating_as::<i32>();
+                    match self.move_cursor(delta) {
+                        Ok(delta) | Err(delta) => {
+                            if chars > 0 {
+                                handler.whitespace("", 1, delta.saturating_as())?;
+                            } else {
+                                handler.move_cursor(delta)?;
+                                handler.whitespace("", 1, delta.abs().saturating_as())?;
+                                handler.move_cursor(delta)?;
                             }
                         }
+                    }
+                }
 
-                        AnsiSequence::CursorForward(n) => {
-                            // Cursor movement can't rely on the text, as it's permitted
-                            // to move the cursor outside of the current line.
-                            // Example:
-                            // (| denotes the cursor, [ and ] are the limits of the line):
-                            // [Some text|    ]
-                            // Cursor forward 2 characters
-                            // [Some text  |  ]
-                            let delta = (n * handler.measure(" ")).saturating_as();
-                            match self.move_cursor(delta) {
-                                Ok(delta) | Err(delta) => {
-                                    handler.whitespace("", 1, delta.saturating_as())?;
-                                }
-                            }
-                        }
-
-                        AnsiSequence::CursorBackward(n) => {
-                            // The above poses an issue with variable-width fonts.
-                            // If cursor movement ignores the variable width, the cursor
-                            // will be placed in positions other than glyph boundaries.
-                            let delta = -(n * handler.measure(" ")).saturating_as::<i32>();
-                            match self.move_cursor(delta) {
-                                Ok(delta) | Err(delta) => {
-                                    handler.move_cursor(delta)?;
-                                    handler.whitespace("", 1, delta.abs().saturating_as())?;
-                                    handler.move_cursor(delta)?;
-                                }
-                            }
-                        }
-
-                        _ => {
-                            // ignore for now
+                Token::MoveCursor {
+                    chars,
+                    draw_background: false,
+                } => {
+                    let delta = chars * handler.measure(" ").saturating_as::<i32>();
+                    match self.move_cursor(delta) {
+                        Ok(delta) | Err(delta) => {
+                            handler.move_cursor(delta)?;
                         }
                     }
                 }
@@ -478,7 +457,7 @@ where
 }
 
 #[cfg(test)]
-mod test {
+pub(crate) mod test {
     use std::convert::Infallible;
 
     use super::*;
@@ -497,7 +476,7 @@ mod test {
     };
 
     #[derive(PartialEq, Eq, Debug)]
-    pub(super) enum RenderElement<C: PixelColor> {
+    pub enum RenderElement<C: PixelColor> {
         Space(u32, bool),
         String(String, u32),
         MoveCursor(i32),
@@ -555,7 +534,6 @@ mod test {
             Ok(())
         }
 
-        #[cfg(feature = "ansi")]
         fn change_text_style(
             &mut self,
             change: ChangeTextStyle<Self::Color>,
@@ -566,7 +544,7 @@ mod test {
     }
 
     #[track_caller]
-    pub(super) fn assert_line_elements<'a, M>(
+    pub(crate) fn assert_line_elements<'a, M>(
         parser: &mut Parser<'a, Rgb888>,
         max_chars: u32,
         elements: &[RenderElement<Rgb888>],
@@ -776,65 +754,5 @@ mod test {
         let mw = PluginWrapper::new(NoPlugin::<Rgb888>::new());
 
         assert_line_elements(&mut parser, 2, &[RenderElement::string("So", 12)], &mw);
-    }
-}
-
-#[cfg(all(test, feature = "ansi"))]
-mod ansi_parser_tests {
-    use super::{
-        test::{assert_line_elements, RenderElement},
-        *,
-    };
-    use crate::plugin::{NoPlugin, PluginWrapper};
-
-    use embedded_graphics::pixelcolor::Rgb888;
-
-    #[test]
-    fn colors() {
-        let mut parser = Parser::parse("Lorem \x1b[92mIpsum");
-        let mw = PluginWrapper::new(NoPlugin::<Rgb888>::new());
-
-        assert_line_elements(
-            &mut parser,
-            100,
-            &[
-                RenderElement::string("Lorem", 30),
-                RenderElement::Space(6, true),
-                RenderElement::ChangeTextStyle(ChangeTextStyle::TextColor(Some(Rgb888::new(
-                    22, 198, 12,
-                )))),
-                RenderElement::string("Ipsum", 30),
-            ],
-            &mw,
-        );
-    }
-
-    #[test]
-    fn ansi_code_does_not_break_word() {
-        let mut parser = Parser::parse("Lorem foo\x1b[92mbarum");
-        let mw = PluginWrapper::new(NoPlugin::<Rgb888>::new());
-
-        assert_line_elements(
-            &mut parser,
-            8,
-            &[
-                RenderElement::string("Lorem", 30),
-                RenderElement::Space(6, false),
-            ],
-            &mw,
-        );
-
-        assert_line_elements(
-            &mut parser,
-            8,
-            &[
-                RenderElement::string("foo", 18),
-                RenderElement::ChangeTextStyle(ChangeTextStyle::TextColor(Some(Rgb888::new(
-                    22, 198, 12,
-                )))),
-                RenderElement::string("barum", 30),
-            ],
-            &mw,
-        );
     }
 }